% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_dose_selector.R
\name{as_tibble.derived_dose_selector}
\alias{as_tibble.derived_dose_selector}
\title{#' @export
print.derived_dose_selector <- function(x, ...) {
  print(x$parent, ...)
}}
\usage{
\method{as_tibble}{derived_dose_selector}(x, ...)
}
\description{
#' @export
print.derived_dose_selector <- function(x, ...) {
  print(x$parent, ...)
}
}
