% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{as_tibble.selector}
\alias{as_tibble.selector}
\title{Cast \code{dose_selector} object to \code{\link[tibble]{tibble}}.}
\usage{
\method{as_tibble}{selector}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dose_selector}.}

\item{...}{Extra args passed onwards.}
}
\value{
Object of class \code{\link[tibble]{tibble}}
}
\description{
Cast \code{dose_selector} object to \code{\link[tibble]{tibble}}.
}
