% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dose_selector_consistency.R
\name{check_dose_selector_consistency}
\alias{check_dose_selector_consistency}
\title{Check the consistency of a dose_selector instance}
\usage{
check_dose_selector_consistency(x)
}
\arguments{
\item{x}{dose_selector}
}
\description{
Check the consistency of a dose_selector instance
}
\examples{
boin_fitter <- get_boin(num_doses = 5, target = 0.3)
x <- fit(boin_fitter, "1NNN")
check_dose_selector_consistency(x)
}
