% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simulations_consistency.R
\name{check_simulations_consistency}
\alias{check_simulations_consistency}
\title{Check the consistency of a dose_selector instance}
\usage{
check_simulations_consistency(x)
}
\arguments{
\item{x}{simulations object}
}
\description{
Check the consistency of a dose_selector instance
}
