% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cohort}
\alias{cohort}
\title{Cohort numbers of evaluated patients.}
\usage{
cohort(selector, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
an integer vector
}
\description{
Get a vector of integers that reflect the cohorts to which the evaluated
patients belong.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% cohort()
}
