% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cohorts_of_n}
\alias{cohorts_of_n}
\title{Sample times between patient arrivals using the exponential distribution.}
\usage{
cohorts_of_n(n = 3, mean_time_delta = 1)
}
\arguments{
\item{n}{integer, sample arrival times for this many patients.}

\item{mean_time_delta}{the average gap between patient arrival times. I.e.
the reciprocal of the rate parameter in an Exponential distribution.}
}
\value{
\code{data.frame} with column time_delta containing durations of time
between patient arrivals.
}
\description{
Sample times between patient arrivals using the exponential distribution.
}
\examples{
cohorts_of_n()
cohorts_of_n(n = 10, mean_time_delta = 5)
}
