% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_selector.R
\name{combo_selector}
\alias{combo_selector}
\title{Dose selector for combinations of treatments}
\usage{
combo_selector()
}
\description{
TODO
}
\examples{
# TODO
}
\seealso{
\code{\link{selector}}
#' @export
}
