% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plot.R
\name{convergence_plot}
\alias{convergence_plot}
\title{Plot the convergence processes from a collection of simulations.}
\usage{
convergence_plot(x, ...)
}
\arguments{
\item{x}{object of type \code{\link{simulations_collection}}}
}
\value{
a ggplot2 plot
}
\description{
Plot the convergence processes from a collection of simulations.
}
\examples{
\dontrun{
# See ? simulate_compare
}
}
