% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_patient_sample.R
\name{CorrelatedPatientSample}
\alias{CorrelatedPatientSample}
\title{A sample of patients that experience correlated events in simulations.}
\description{
Class to house the latent random variables that govern toxicity and efficacy
events in patients. Instances of this class can be used in simulation-like
tasks to effectively use the same simulated individuals in different designs,
thus supporting reduced Monte Carlo error and more efficient comparison. This
class differs from \code{\link{PatientSample}} in that the latent variables
that underly efficacy and toxicity events, and therefore those events
themselves, are correlated, e.g. for positive association, a patient that
experiences toxicity has increased probability of experiencing efficacy too.
The extent to which the events are correlated is controlled by tau. TODO
}
\references{
Sweeting, M., Slade, D., Jackson, D., & Brock, K. (2023).
Potential outcome simulation for efficient head-to-head comparison of
adaptive dose-finding designs. Preprint.
}
\section{Super class}{
\code{\link[escalation:PatientSample]{escalation::PatientSample}} -> \code{CorrelatedPatientSample}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`)\cr}

\item{\code{tau}}{(`numeric(1)`)\cr}

\item{\code{prob_event}}{(`numeric(41)`)\cr}

\item{\code{xi}}{(`numeric(41)`)\cr}

\item{\code{theta}}{(`numeric(num_patients)`)\cr}

\item{\code{can_grow}}{(`logical(1)`)\cr}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CorrelatedPatientSample-new}{\code{CorrelatedPatientSample$new()}}
\item \href{#method-CorrelatedPatientSample-set_theta}{\code{CorrelatedPatientSample$set_theta()}}
\item \href{#method-CorrelatedPatientSample-expand_to}{\code{CorrelatedPatientSample$expand_to()}}
\item \href{#method-CorrelatedPatientSample-get_theta}{\code{CorrelatedPatientSample$get_theta()}}
\item \href{#method-CorrelatedPatientSample-get_patient_tox}{\code{CorrelatedPatientSample$get_patient_tox()}}
\item \href{#method-CorrelatedPatientSample-get_patient_eff}{\code{CorrelatedPatientSample$get_patient_eff()}}
\item \href{#method-CorrelatedPatientSample-clone}{\code{CorrelatedPatientSample$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="escalation" data-topic="PatientSample" data-id="get_eff_u"><a href='../../escalation/html/PatientSample.html#method-PatientSample-get_eff_u'><code>escalation::PatientSample$get_eff_u()</code></a></span></li>
<li><span class="pkg-link" data-pkg="escalation" data-topic="PatientSample" data-id="get_tox_u"><a href='../../escalation/html/PatientSample.html#method-PatientSample-get_tox_u'><code>escalation::PatientSample$get_tox_u()</code></a></span></li>
<li><span class="pkg-link" data-pkg="escalation" data-topic="PatientSample" data-id="set_eff_and_tox"><a href='../../escalation/html/PatientSample.html#method-PatientSample-set_eff_and_tox'><code>escalation::PatientSample$set_eff_and_tox()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-new"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-new}{}}}
\subsection{Method \code{new()}}{
Creator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$new(num_patients = 0, tau = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`).}

\item{\code{tau}}{(`integer(1)`) standard deviation of normal latent variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[PatientSample].
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-set_theta"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-set_theta}{}}}
\subsection{Method \code{set_theta()}}{
Set the theta latent variables that govern occurrence of
toxicity and efficacy events. By default, instances of this class
automatically grow these latent variables to accommodate arbitrarily high
sample sizes. However, when you set these latent variables manually via
this function, you override the ability of the class to self-manage, so
its ability to grow is turned off by setting the internal variable
\code{self$can_grow <- FALSE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$set_theta(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{(`numeric()`).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-expand_to"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-expand_to}{}}}
\subsection{Method \code{expand_to()}}{
Expand sample to size at least num_patients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$expand_to(num_patients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`).
#' @description
#' Get toxicity latent variable for patient i
#'
#' @param i (`integer(1)`) patient index
get_tox_u = function(i) {
  if(i > self$num_patients) {
    self$expand_to(num_patients = i)
  }
  return(self$tox_u[i])
},}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-get_theta"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-get_theta}{}}}
\subsection{Method \code{get_theta()}}{
Get theta latent variable for patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$get_theta(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-get_patient_tox"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-get_patient_tox}{}}}
\subsection{Method \code{get_patient_tox()}}{
Get 0 or 1 event marker for whether toxicity occurred in patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$get_patient_tox(i, prob_tox)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}

\item{\code{prob_tox}}{(`numeric(1)`) probability of toxicity
#' @description
#' Get efficacy latent variable for patient i
#'
#' @param i (`integer(1)`) patient index
get_eff_u = function(i) {
  if(i > self$num_patients) {
    self$expand_to(num_patients = i)
  }
  return(self$eff_u[i])
},}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-get_patient_eff"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-get_patient_eff}{}}}
\subsection{Method \code{get_patient_eff()}}{
Get 0 or 1 event marker for whether efficacy occurred in patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$get_patient_eff(i, prob_eff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}

\item{\code{prob_eff}}{(`numeric(1)`) probability of efficacy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatedPatientSample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatedPatientSample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatedPatientSample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
