% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crystallised_dose_paths.R
\name{crystallised_dose_paths}
\alias{crystallised_dose_paths}
\title{Dose-paths with probabilities attached.}
\usage{
crystallised_dose_paths(dose_paths, true_prob_tox, terminal_nodes)
}
\arguments{
\item{dose_paths}{Object of type \code{\link{dose_paths}}}

\item{true_prob_tox}{vector of probabilities}

\item{terminal_nodes}{tibble of terminal nodes on the dose-paths}
}
\value{
An object of type crystallised_dose_paths
}
\description{
\code{\link{dose_paths}} reflect all possible paths a dose-finding trial may
take. When the probability of those paths is calculated using an assumed set
of true dose-event probabilities, in this package those paths are said to be
crysallised. Once crystallised, operating charactersitics can be calculated.
}
\examples{
1 == 1  # TODO
}
