% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_paths.R
\name{dose_paths}
\alias{dose_paths}
\title{Dose pathways}
\usage{
dose_paths()
}
\description{
A dose-escalation design exists to select doses in response to observed
outcomes. The entire space of possible responses can be calculated to show
the behaviour of a design in response to all feasible outcomes. The
\code{\link{get_dose_paths}} function performs that task and returns an
instance of this object.
}
\examples{

# Calculate dose-paths for the 3+3 design:
paths <- get_three_plus_three(num_doses = 5) \%>\%
  get_dose_paths(cohort_sizes = c(3, 3))
}
\seealso{
\code{\link{selector}}
}
