% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dose_paths_function}
\alias{dose_paths_function}
\title{Get function for calculating dose pathways.}
\usage{
dose_paths_function(selector_factory)
}
\arguments{
\item{selector_factory}{Object of type \code{\link{selector_factory}}.}
}
\value{
A function.
}
\description{
This function does not need to be called by users. It is used internally.
}
