% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dose_string_to_vector}
\alias{dose_string_to_vector}
\title{Go from a single multi-treatment dose string to a vector of dose-indices}
\usage{
dose_string_to_vector(x)
}
\arguments{
\item{x}{a multi-treatment dose string like "1.3" or "3.2"}
}
\value{
a vector of integer dose-indices
}
\description{
Go from a single multi-treatment dose string to a vector of dose-indices
}
\examples{
dose_string_to_vector("1.3")
}
