% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{dose_strings}
\alias{dose_strings}
\title{Dose strings}
\usage{
dose_strings(x, ...)
}
\arguments{
\item{x}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
a character vector
}
\description{
Get strings representing the doses under investigation, congruent to
dose_indices(x)
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% dose_strings()
}
