% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dose_vector_to_string}
\alias{dose_vector_to_string}
\title{Go from a single multi-treatment vector of dose-indices to a dose string}
\usage{
dose_vector_to_string(x)
}
\arguments{
\item{x}{vector of integer dose indices}
}
\value{
character
}
\description{
Go from a single multi-treatment vector of dose-indices to a dose string
}
\examples{
dose_vector_to_string(as.integer(c(1, 3))) # The indices must be integers!
}
