% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_dose_combo_indices}
\alias{get_dose_combo_indices}
\title{Get all combinations of dose indices}
\usage{
get_dose_combo_indices(num_doses)
}
\arguments{
\item{num_doses}{integer vector of number of doses for length(num_doses)
treatments}
}
\value{
a list, each element being an integer vector
}
\description{
Get all combinations of dose indices
}
\examples{
get_dose_combo_indices(num_doses = c(1, 2))
# returns list(c(1, 1), c(1, 2))
}
