% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpi_selector.R
\name{get_tpi}
\alias{get_tpi}
\title{Get an object to fit the TPI dose-finding model.}
\usage{
get_tpi(
  num_doses,
  target,
  k1,
  k2,
  exclusion_certainty,
  alpha = 0.005,
  beta = 0.005,
  ...
)
}
\arguments{
\item{num_doses}{Number of doses under investigation.}

\item{target}{We seek a dose with this probability of toxicity.}

\item{k1}{The K1 parameter in TPI determines the lower bound of the
equivalence interval. See Details.}

\item{k2}{The K2 parameter in TPI determines the upper bound of the
equivalence interval. See Details.}

\item{exclusion_certainty}{Numeric, threshold posterior certainty required to
exclude a dose for being excessively toxic. The authors discuss values in the
range 0.7 - 0.95. Set to a value > 1 to suppress the dose exclusion
mechanism.}

\item{alpha}{First shape parameter of the beta prior distribution on the
probability of toxicity.}

\item{beta}{Second shape parameter of the beta prior distribution on the
probability of toxicity.}

\item{...}{Extra args are passed onwards.}
}
\value{
an object of type \code{\link{selector_factory}} that can fit the
TPI model to outcomes.
}
\description{
Get an object to fit the TPI dose-finding model.
}
\examples{
target <- 0.25
model1 <- get_tpi(num_doses = 5, target = target, k1 = 1, k2 = 1.5,
  exclusion_certainty = 0.95)

outcomes <- '1NNN 2NTN'
model1 \%>\% fit(outcomes) \%>\% recommended_dose()

}
\references{
Ji, Y., Li, Y., & Bekele, B. N. (2007).
Dose-finding in phase I clinical trials based on toxicity probability
intervals.
Clinical Trials, 4(3), 235–244. https://doi.org/10.1177/1740774507079442

Ji, Y., & Yang, S. (2017).
On the Interval-Based Dose-Finding Designs, 1–26.
Retrieved from http://arxiv.org/abs/1706.03277
}
