% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{is_randomising}
\alias{is_randomising}
\title{Is this selector currently randomly allocating doses?}
\usage{
is_randomising(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{selector}}}

\item{...}{arguments passed to other methods}
}
\value{
a logical value
}
\description{
Get the percentage of patients evaluated at each dose under investigation.
}
\examples{
outcomes <- '1NNN 2NTN'
fit <- get_random_selector(prob_select = c(0.1, 0.6, 0.3)) \%>\%
  fit(outcomes)
fit \%>\% is_randomising()
}
