% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_collection.R
\name{make_simulations_collection}
\alias{make_simulations_collection}
\title{Make an instance of type \code{simulations_collection}}
\usage{
make_simulations_collection(sim_map)
}
\arguments{
\item{sim_map}{list, character -> \class{\link{simulations}} object}
}
\value{
object of class \code{simulations_collection}, inheriting from list
}
\description{
Make an instance of type \code{simulations_collection}
}
\examples{
# TODO
}
