% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boin_comb_selector.R
\name{mean_prob_tox.boin_comb_selector}
\alias{mean_prob_tox.boin_comb_selector}
\title{#' @export
tox_at_dose.boin_comb_selector <- function(x, ...) {
  return(x$df_c$tox)
}}
\usage{
\method{mean_prob_tox}{boin_comb_selector}(x, ...)
}
\description{
#' @export
tox_at_dose.boin_comb_selector <- function(x, ...) {
  return(x$df_c$tox)
}
}
