% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{num_doses}
\alias{num_doses}
\title{Number of doses.}
\usage{
num_doses(selector, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
integer
}
\description{
Get the number of doses under investigation in a dose-finding trial.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% num_doses()
}
