% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcrm_selector.R
\name{num_patients.dfcrm_selector}
\alias{num_patients.dfcrm_selector}
\title{Title}
\usage{
\method{num_patients}{dfcrm_selector}(selector, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
