% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{num_patients}
\alias{num_patients}
\title{Number of patients evaluated.}
\usage{
num_patients(selector, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
integer
}
\description{
Get the number of patients evaluated in a dose-finding trial.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton, target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% num_patients()
}
