% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{num_tox}
\alias{num_tox}
\title{Total number of toxicities seen.}
\usage{
num_tox(selector, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
integer
}
\description{
Get the number of toxicities seen in a dose-finding trial.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% num_tox()
}
