% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_sample.R
\name{PatientSample}
\alias{PatientSample}
\title{Patient tolerance samples for simulations}
\description{
Class to house the latent random variables that govern tox and eff events in
simulations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`)\cr}

\item{\code{tox_u}}{(`numeric(num_patients)`)\cr}

\item{\code{eff_u}}{(`numeric(num_patients)`)\cr}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PatientSample-new}{\code{PatientSample$new()}}
\item \href{#method-PatientSample-expand_to}{\code{PatientSample$expand_to()}}
\item \href{#method-PatientSample-get_tox_u}{\code{PatientSample$get_tox_u()}}
\item \href{#method-PatientSample-get_patient_tox}{\code{PatientSample$get_patient_tox()}}
\item \href{#method-PatientSample-get_eff_u}{\code{PatientSample$get_eff_u()}}
\item \href{#method-PatientSample-get_patient_eff}{\code{PatientSample$get_patient_eff()}}
\item \href{#method-PatientSample-clone}{\code{PatientSample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-new"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-new}{}}}
\subsection{Method \code{new()}}{
Creator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$new(num_patients = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[PatientSample].
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-expand_to"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-expand_to}{}}}
\subsection{Method \code{expand_to()}}{
Expand sample to size at least num_patients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$expand_to(num_patients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_patients}}{(`integer(1)`).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-get_tox_u"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-get_tox_u}{}}}
\subsection{Method \code{get_tox_u()}}{
Get toxicity latent variable for patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$get_tox_u(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-get_patient_tox"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-get_patient_tox}{}}}
\subsection{Method \code{get_patient_tox()}}{
Get 0 or 1 event marker for whether toxicity occurred in patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$get_patient_tox(i, prob_tox)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}

\item{\code{prob_tox}}{(`numeric(1)`) probability of toxicity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-get_eff_u"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-get_eff_u}{}}}
\subsection{Method \code{get_eff_u()}}{
Get efficacy latent variable for patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$get_eff_u(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-get_patient_eff"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-get_patient_eff}{}}}
\subsection{Method \code{get_patient_eff()}}{
Get 0 or 1 event marker for whether efficacy occurred in patient i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$get_patient_eff(i, prob_eff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{(`integer(1)`) patient index}

\item{\code{prob_eff}}{(`numeric(1)`) probability of efficacy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PatientSample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PatientSample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PatientSample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
