% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{recommended_dose}
\alias{recommended_dose}
\title{Recommended dose for next patient or cohort.}
\usage{
recommended_dose(selector, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
integer
}
\description{
Get the dose recommended for the next patient or cohort in a dose-finding
trial.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton, target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% recommended_dose()
}
