% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trials.R
\name{simulate_trials}
\alias{simulate_trials}
\title{Simulate clinical trials.}
\usage{
simulate_trials(selector_factory, num_sims, true_prob_tox, ...)
}
\arguments{
\item{selector_factory}{Object of type \code{\link{selector_factory}}.}

\item{num_sims}{integer, number of trial iterations to simulate.}

\item{true_prob_tox}{numeric vector of toxicity probabilities}

\item{...}{Extra args are passed onwards.}
}
\value{
Object of type \code{\link{simulations}}.
}
\description{
Simulate clinical trials.
}
\examples{
# Simulate performance of the 3+3 design:
true_prob_tox <- c(0.12, 0.27, 0.44, 0.53, 0.57)
sims <- get_three_plus_three(num_doses = 5) \%>\%
  simulate_trials(num_sims = 50, true_prob_tox = true_prob_tox)
}
