% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_collection.R
\name{simulations_collection}
\alias{simulations_collection}
\title{Make an instance of type \code{simulations_collection}}
\usage{
simulations_collection(sim_map)
}
\arguments{
\item{sim_map}{list, character -> \code{\link{simulations}} object}
}
\value{
object of class \code{simulations_collection}, inheriting from list
}
\description{
This object can be cast to a tibble with \code{as_tibble} to generate useful
pairwise comparisons of the probability of recommending each dose for each
pair of designs investigated. See
\code{\link{as_tibble.simulations_collection}} for a description.
}
\references{
Sweeting, M., Slade, D., Jackson, D., & Brock, K. (2024).
Potential outcome simulation for efficient head-to-head comparison of
adaptive dose-finding designs. arXiv preprint arXiv:2402.15460
}
