% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_sims_vert.R
\name{stack_sims_vert}
\alias{stack_sims_vert}
\title{Stack \code{\link{simulations_collection}} results vertically}
\usage{
stack_sims_vert(sim_map, target_dose = NULL, alpha = 0.05)
}
\arguments{
\item{sim_map}{object of type \code{\link{simulations_collection}}}

\item{target_dose}{optional integer vector, the dose of interest. All doses
are analysed if omitted, which is the default.}

\item{alpha}{confidence level for asymptotic normal confidence intervals. The
default value is 0.05 to get 95% confidence intervals.}
}
\value{
\code{data.frame}
}
\description{
Stack \code{\link{simulations_collection}} results vertically
}
\examples{
# TODO
}
