% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{weight}
\alias{weight}
\title{Outcome weights.}
\usage{
weight(x, ...)
}
\arguments{
\item{x}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
a numerical vector
}
\description{
Get a vector of the weights attached to outcomes for evaluated patients.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% weight()
}
