% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_collection.R
\name{as_tibble.simulations_collection}
\alias{as_tibble.simulations_collection}
\title{Compare the probabilities of dose selection in several simulation sets.}
\usage{
\method{as_tibble}{simulations_collection}(sim_map, target_dose = NULL, alpha = 0.05)
}
\arguments{
\item{sim_map}{object of type \code{\link{simulations_collection}}}

\item{target_dose}{optional integer vector, the dose of interest. All doses
are analysed if omitted, which is the default.}

\item{alpha}{confidence level for asymptotic normal confidence intervals. The
default value is 0.05 to get 95% confidence intervals.}
}
\value{
\code{data.frame}
}
\description{
Normal approximation in derivation of confidence intervals of differences in
probability of selection. Details are in Sweeting et al.
}
\examples{
# TODO

}
\references{
# TODO
}
