% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities.R
\name{calculate_probabilities}
\alias{calculate_probabilities}
\title{Calculate dose-path probabilities}
\usage{
calculate_probabilities(dose_paths, true_prob_tox)
}
\arguments{
\item{dose_paths}{Object of type \code{\link{dose_paths}}}

\item{true_prob_tox}{Numeric vector, true probability of toxicity.}
}
\description{
Crystallise a set of \code{\link{dose_paths}} with probabilities to calculate
how likely each path is. Once probabilised in this way, the probabilities of
the terminal nodes in this set of paths will sum to 1. This allows users to
calculate operating characteristics.
}
\examples{
1 == 1 # TODO
}
\seealso{
\code{\link{dose_paths}}
}
