% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{doses_given}
\alias{doses_given}
\title{Doses given to patients.}
\usage{
doses_given(x, ...)
}
\arguments{
\item{x}{Object of type \code{\link{selector}}.}

\item{...}{Extra args are passed onwards.}
}
\value{
an integer vector
}
\description{
Get a vector of the dose-levels that have been administered to patients.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model <- get_dfcrm(skeleton = skeleton, target = target)
fit <- model \%>\% fit('1NNN 2NTN')
fit \%>\% doses_given()
}
