% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_last_dose_to_cohort.R
\name{expand_last_dose_to_cohort}
\alias{expand_last_dose_to_cohort}
\title{Title}
\usage{
expand_last_dose_to_cohort(parent_selector_factory, n)
}
\arguments{
\item{n}{}
}
\description{
Title
}
