% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boin_selector.R
\name{get_boin}
\alias{get_boin}
\title{Get an object to fit the CRM model using the dfcrm package.}
\usage{
get_boin(num_doses, target, ...)
}
\arguments{
\item{num_doses}{Number of doses under investigation.}

\item{target}{We seek a dose with this probability of toxicity.}

\item{...}{Extra args are passed to \code{\link[BOIN]{select.mtd}}.}
}
\value{
an object of type \code{\link{selector_factory}} that can fit the
BOIN model to outcomes.
}
\description{
Get an object to fit the CRM model using the dfcrm package.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model1 <- get_dfcrm(skeleton, target)

outcomes <- '1NNN 2NTN'
model1 \%>\% fit(outcomes) \%>\% recommended_dose()

model2 <- get_dfcrm(skeleton, target, model = 'logistic')
model2 \%>\% fit(outcomes) \%>\% recommended_dose()

}
\references{
TODO
}
