% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boin_comb_selector.R
\name{get_boin_comb}
\alias{get_boin_comb}
\title{Get an object to fit the BOIN COMB model using the BOIN package.}
\usage{
get_boin_comb(num_doses, target, use_stopping_rule = TRUE, ...)
}
\arguments{
\item{num_doses}{integer vector of the number of doses of treatment 1, 2}

\item{target}{We seek a dose with this probability of toxicity.}

\item{use_stopping_rule}{TODO}

\item{...}{Extra args are passed to \code{\link[BOIN]{next.comb}}.}
}
\value{
an object of type \code{\link{selector_factory}} that can fit the
BOIN COMB model to outcomes.
}
\description{
Get an object to fit the BOIN COMB model using the BOIN package.
}
\examples{
# TODO

}
\references{
Lin, R., & Yin, G. (2017).
Bayesian optimal interval design for dose finding in drug-combination trials.
Statistical methods in medical research, 26(5), 2155-2167.
}
