% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{prob_tox_exceeds}
\alias{prob_tox_exceeds}
\title{Probability that the toxicity rate exceeds some threshold.}
\usage{
prob_tox_exceeds(selector, threshold, ...)
}
\arguments{
\item{selector}{Object of type \code{\link{selector}}}

\item{threshold}{Probability that toxicity rate exceeds what?}

\item{...}{arguments passed to other methods}
}
\value{
numerical vector of probabilities
}
\description{
Get the probability that the toxicity rate at each dose exceeds some
threshold.
}
\examples{
1==1
}
