% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_at_n_selector.R
\name{stop_at_n}
\alias{stop_at_n}
\title{Stop when there are n patients in total.}
\usage{
stop_at_n(parent_selector_factory, n)
}
\arguments{
\item{parent_selector_factory}{Object of type \code{\link{selector_factory}}.}

\item{n}{Stop when there are this many patients.}
}
\value{
an object of type \code{\link{selector_factory}} that can fit a
dose-finding model to outcomes.
}
\description{
This method adds together the number of patients treated at all doses and
stops a dose-finding trial when there are at least n patients in total.
}
\examples{
skeleton <- c(0.05, 0.1, 0.25, 0.4, 0.6)
target <- 0.25
model1 <- get_dfcrm(skeleton = skeleton, target = target) \%>\%
  stop_at_n(n = 15)

fit1 <- model1 \%>\% fit('1NNN 2NTN 2TNN 2NNN')
fit1 \%>\% recommended_dose()
fit1 \%>\% continue()
fit2 <- model1 \%>\% fit('1NNN 2NTN 2TNN 2NNN 2NTT')
fit2 \%>\% recommended_dose()
fit2 \%>\% continue()
}
